refactor(evaluation): move funçoes para a controller correta
@eluvya

103
